+++ FeishuSpeech/Models/StreamingSpeechModels.swift
@@ -31,24 +31,119 @@ nonisolated struct StreamingRetryPolicy: Sendable {
 }
 
 nonisolated struct StreamingDrainPolicy: Equatable, Sendable {
-    let operationTimeoutNanoseconds: UInt64
+    let factoryTimeoutNanoseconds: UInt64
+    let packetTimeoutNanoseconds: UInt64
+    let finishTimeoutNanoseconds: UInt64
     let postReleaseDrainTimeoutNanoseconds: UInt64
+    let urlSessionFactorySliceNanoseconds: UInt64
+    let directFactorySliceNanoseconds: UInt64
+    let urlSessionPacketSliceNanoseconds: UInt64
+    let directPacketSliceNanoseconds: UInt64
+    let urlSessionFinishSliceNanoseconds: UInt64
+    let directFinishSliceNanoseconds: UInt64
+    let sliceSlackNanoseconds: UInt64
+
+    var operationTimeoutNanoseconds: UInt64 {
+        packetTimeoutNanoseconds
+    }
 
     init(
-        operationTimeoutNanoseconds: UInt64 = 30_000_000_000,
+        factoryTimeoutNanoseconds: UInt64 = 18_000_000_000,
+        packetTimeoutNanoseconds: UInt64 = 30_000_000_000,
+        finishTimeoutNanoseconds: UInt64 = 45_000_000_000,
+        postReleaseDrainTimeoutNanoseconds: UInt64 = 60_000_000_000,
+        urlSessionFactorySliceNanoseconds: UInt64 = 8_000_000_000,
+        directFactorySliceNanoseconds: UInt64 = 7_000_000_000,
+        urlSessionPacketSliceNanoseconds: UInt64 = 14_000_000_000,
+        directPacketSliceNanoseconds: UInt64 = 14_000_000_000,
+        urlSessionFinishSliceNanoseconds: UInt64 = 15_000_000_000,
+        directFinishSliceNanoseconds: UInt64 = 15_000_000_000,
+        sliceSlackNanoseconds: UInt64 = 1_000_000_000
+    ) {
+        precondition(factoryTimeoutNanoseconds > 0)
+        precondition(packetTimeoutNanoseconds > 0)
+        precondition(finishTimeoutNanoseconds > 0)
+        precondition(postReleaseDrainTimeoutNanoseconds > 0)
+        precondition(sliceSlackNanoseconds >= 500_000_000 || sliceSlackNanoseconds == 1)
+        precondition(
+            urlSessionFactorySliceNanoseconds &+ directFactorySliceNanoseconds &+ sliceSlackNanoseconds
+                < factoryTimeoutNanoseconds
+        )
+        precondition(
+            urlSessionPacketSliceNanoseconds &+ directPacketSliceNanoseconds &+ sliceSlackNanoseconds
+                < packetTimeoutNanoseconds
+        )
+        precondition(
+            urlSessionFinishSliceNanoseconds &+ directFinishSliceNanoseconds &+ sliceSlackNanoseconds
+                < finishTimeoutNanoseconds
+        )
+        self.factoryTimeoutNanoseconds = factoryTimeoutNanoseconds
+        self.packetTimeoutNanoseconds = packetTimeoutNanoseconds
+        self.finishTimeoutNanoseconds = finishTimeoutNanoseconds
+        self.postReleaseDrainTimeoutNanoseconds = postReleaseDrainTimeoutNanoseconds
+        self.urlSessionFactorySliceNanoseconds = urlSessionFactorySliceNanoseconds
+        self.directFactorySliceNanoseconds = directFactorySliceNanoseconds
+        self.urlSessionPacketSliceNanoseconds = urlSessionPacketSliceNanoseconds
+        self.directPacketSliceNanoseconds = directPacketSliceNanoseconds
+        self.urlSessionFinishSliceNanoseconds = urlSessionFinishSliceNanoseconds
+        self.directFinishSliceNanoseconds = directFinishSliceNanoseconds
+        self.sliceSlackNanoseconds = sliceSlackNanoseconds
+    }
+
+    init(
+        operationTimeoutNanoseconds: UInt64,
         postReleaseDrainTimeoutNanoseconds: UInt64 = 60_000_000_000
     ) {
-        precondition(operationTimeoutNanoseconds > 0)
-        precondition(postReleaseDrainTimeoutNanoseconds > 0)
-        self.operationTimeoutNanoseconds = operationTimeoutNanoseconds
-        self.postReleaseDrainTimeoutNanoseconds = postReleaseDrainTimeoutNanoseconds
+        let outer = operationTimeoutNanoseconds
+        self.init(
+            factoryTimeoutNanoseconds: outer,
+            packetTimeoutNanoseconds: outer,
+            finishTimeoutNanoseconds: outer,
+            postReleaseDrainTimeoutNanoseconds: postReleaseDrainTimeoutNanoseconds,
+            urlSessionFactorySliceNanoseconds: 1,
+            directFactorySliceNanoseconds: 1,
+            urlSessionPacketSliceNanoseconds: 1,
+            directPacketSliceNanoseconds: 1,
+            urlSessionFinishSliceNanoseconds: 1,
+            directFinishSliceNanoseconds: 1,
+            sliceSlackNanoseconds: 1
+        )
     }
 
     func operationTimeout(remainingDrainNanoseconds: UInt64?) -> UInt64 {
-        guard let remainingDrainNanoseconds else {
-            return operationTimeoutNanoseconds
+        operationTimeout(for: "packet", remainingDrainNanoseconds: remainingDrainNanoseconds)
+    }
+
+    func operationTimeout(
+        for operation: String,
+        remainingDrainNanoseconds: UInt64?
+    ) -> UInt64 {
+        let outer: UInt64
+        switch operation {
+        case "factory":
+            outer = factoryTimeoutNanoseconds
+        case "finish":
+            outer = finishTimeoutNanoseconds
+        default:
+            outer = packetTimeoutNanoseconds
         }
-        return min(operationTimeoutNanoseconds, remainingDrainNanoseconds)
+        guard let remainingDrainNanoseconds else {
+            return outer
+        }
+        return min(outer, remainingDrainNanoseconds)
+    }
+
+    func urlSessionSliceNanoseconds(for phase: AttemptHTTPPhase) -> UInt64 {
+        switch phase {
+        case .factoryToken:
+            return urlSessionFactorySliceNanoseconds
+        case .packet:
+            return urlSessionPacketSliceNanoseconds
+        case .finish:
+            return urlSessionFinishSliceNanoseconds
+        case .abort:
+            return 1_000_000_000
+        }
     }
 
     func retryDelay(
@@ -60,6 +155,18 @@ nonisolated struct StreamingDrainPolicy: Equatable, Sendable {
         }
         return min(requestedNanoseconds, remainingDrainNanoseconds)
     }
+}
+
+nonisolated enum AttemptHTTPPhase: Equatable, Sendable {
+    case factoryToken
+    case packet
+    case finish
+    case abort
+}
+
+nonisolated struct AttemptHTTPRequest: Sendable {
+    let request: URLRequest
+    let phase: AttemptHTTPPhase
 }
 
 nonisolated struct AudioIngressConfiguration: Equatable, Sendable {